pyproject.toml
[project]
name = "enterprise-rag-haystack"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "httpx>=0.27",

  "haystack-ai>=2.0.0",
  "qdrant-client>=1.9.0",
  "sentence-transformers>=2.6.0",

  # Gemini (Google GenAI SDK)
  "google-genai>=1.0.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]